# repo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CampusFit API
  description: >-
    CampusFit platform API covering authentication, fitness tracking, study planning,
    inbound receipt management, master data administration, notifications, data export,
    and reporting.
  version: 1.0.0
  contact:
    name: CampusFit Team
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/study/export:
    get:
      tags: [Study]
      summary: Export study data as JSON
      operationId: exportStudyData
      responses:
        '200':
          description: Study data JSON payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseStudyExportData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/study/import:
    post:
      tags: [Study]
      summary: Import study data from JSON
      operationId: importStudyData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyExportData'
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── Inbound ────────────────────────────────────────────────────────
...
  /api/admin/majors:
    get:
      tags: [Admin]
      summary: List all majors
      operationId: listMajors
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of majors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseMasterDataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Admin]
      summary: Create a major
      operationId: createMajor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MasterDataRequest'
      responses:
        '201':
          description: Major created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMasterDataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      operationId: listNotifications
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseNotificationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Notifications]
      summary: Create a notification (admin only)
      operationId: createNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - title
                - targetUserIds
              properties:
                type:
                  type: string
                  enum: [ANNOUNCEMENT, REMINDER, FOLLOW_UP]
                title:
                  type: string
                body:
                  type: string
                targetUserIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark a notification as read
      operationId: markNotificationRead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/notifications/{id}/status:
    get:
      tags: [Notifications]
      summary: Get delivery status for a notification
      operationId: getNotificationDeliveryStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Delivery status details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDeliveryStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── Exports / Account ─────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Session token obtained from /api/auth/sign-in
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Page number (zero-based)
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
...
    MasterDataRequest:
      type: object
      required:
        - code
        - name
        - effectiveFrom
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        schoolId:
          type: integer
          format: int64
          nullable: true
        majorId:
          type: integer
          format: int64
          nullable: true
        year:
          type: integer
          nullable: true
        classId:
          type: integer
          format: int64
          nullable: true
        termId:
          type: integer
          format: int64
          nullable: true
        credits:
          type: integer
          nullable: true
    MasterDataResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        schoolId:
          type: integer
          format: int64
          nullable: true
        majorId:
          type: integer
          format: int64
          nullable: true
        year:
          type: integer
          nullable: true
        classId:
          type: integer
          format: int64
          nullable: true
        termId:
          type: integer
          format: int64
          nullable: true
        credits:
          type: integer
          nullable: true
    ApiResponseMasterDataResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MasterDataResponse'
        error:
          type: string
          nullable: true
    PageResponseMasterDataResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MasterDataResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        notificationId:
          type: integer
          format: int64
        type:
          type: string
          enum: [ANNOUNCEMENT, REMINDER, FOLLOW_UP]
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PageResponseNotificationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    DeliveryStatusResponse:
      type: object
      properties:
        targetId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        channel:
          type: string
    ApiResponseDeliveryStatus:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryStatusResponse'
        error:
          type: string
          nullable: true

    # ─── Export Schemas ────────────────────────────────────────────
...
    ApiResponseString:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: string
        error:
          type: string
          nullable: true
...
    StudyExportData:
      type: object
      properties:
        plans:
          type: array
          items:
            type: object
        completions:
          type: array
          items:
            type: object
        forgettingPoints:
          type: array
          items:
            type: object
    ApiResponseStudyExportData:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/StudyExportData'
        error:
          type: string
          nullable: true
...